chore: ErrorLogsProvider and useErrorLogs
